# api.yaml
openapi: 3.0.0
info:
  title: Webservices API
  description: API voor het vak webservices
  version: 0.1.0
servers:
  - url: http://localhost:9000/api
tags:
  - name: User
    description: Everything about users
  - name: Post
    description: Everything about posts
paths:
  /users:
    get:
      security:
        - bearerAuth: []
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - User
      responses:
        '200':    # status code
          description: A JSON array of user object + token
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/{id}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      summary: Returns a specific user.
      tags:
        - User
      responses:
        '200':    # status code
          description: A JSON array of user object + token
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string             
  /users/login:
    post:
      summary: Login a user
      tags:
        - User
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '201':
          description: Created
  /users/register:
    post:
      summary: Sign up a user
      tags:
        - User
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '200':
          description: User Registered

components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    UserRegister:
      type: object
      properties:
        naam:
          type: string
        voornaam:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - naam
        - voornaam
        - email
        - password